Module kube_client::api::entry

API helpers for get-or-create and get-and-modify patterns

Api::entry is the primary entry point for this API.

Structs

OccupiedEntry

A view into a single object that exists

VacantEntry

A view of an object that does not yet exist

Enums

CommitError

Commit errors

CommitValidationError

Pre-commit validation errors

Entry

A view into a single object, with enough context to create or update it