Struct kube_client::api::AttachedProcess

pub struct AttachedProcess { /* private fields */ }

Represents an attached process in a container for attach and exec.

Provides access to stdin, stdout, and stderr if attached.

Use AttachedProcess::join to wait for the process to terminate.

Implementations

impl AttachedProcess

pub fn stdin(&mut self) -> Option<impl AsyncWrite + Unpin>

Async writer to stdin.

let mut stdin_writer = attached.stdin().unwrap();
stdin_writer.write(b"foo\n").await?;

Only available if AttachParams had stdin.

pub fn stdout(&mut self) -> Option<impl AsyncRead + Unpin>

Async reader for stdout outputs.

let mut stdout_reader = attached.stdout().unwrap();
let next_stdout = stdout_reader.read().await?;

Only available if AttachParams had stdout.

pub fn stderr(&mut self) -> Option<impl AsyncRead + Unpin>

Async reader for stderr outputs.

let mut stderr_reader = attached.stderr().unwrap();
let next_stderr = stderr_reader.read().await?;

Only available if AttachParams had stderr.

pub fn abort(&self)

Abort the background task, causing remote command to fail.

pub async fn join(self) -> Result<(), Error>

Waits for the remote command task to complete.

pub fn take_status(&mut self) -> Option<impl Future<Output = Option<Status>>>

Take a future that resolves with any status object or when the sender is dropped.

Returns None if called more than once.

pub fn terminal_size(&mut self) -> Option<Sender<TerminalSize>>

Async writer to change the terminal size

let mut terminal_size_writer = attached.terminal_size().unwrap();
terminal_size_writer.send(TerminalSize{
    height: 100,
    width: 200,
}).await?;

Only available if AttachParams had tty.