Struct kube_client::api::DeleteParams

pub struct DeleteParams {
    pub dry_run: bool,
    pub grace_period_seconds: Option<u32>,
    pub propagation_policy: Option<PropagationPolicy>,
    pub preconditions: Option<Preconditions>,
}

Common query parameters for delete calls

Fields
dry_run: bool

When present, indicates that modifications should not be persisted.

grace_period_seconds: Option<u32>

The duration in seconds before the object should be deleted.

Value must be non-negative integer. The value zero indicates delete immediately. If this value is None, the default grace period for the specified type will be used. Defaults to a per object value if not specified. Zero means delete immediately.

propagation_policy: Option<PropagationPolicy>

Whether or how garbage collection is performed.

The default policy is decided by the existing finalizer set in metadata.finalizers, and the resource-specific default policy.

preconditions: Option<Preconditions>

Condtions that must be fulfilled before a deletion is carried out

If not possible, a 409 Conflict status will be returned.

Implementations

impl DeleteParams

pub fn background() -> DeleteParams

Construct DeleteParams with PropagationPolicy::Background.

This allows the garbage collector to delete the dependents in the background.

pub fn foreground() -> DeleteParams

Construct DeleteParams with PropagationPolicy::Foreground.

This is a cascading policy that deletes all dependents in the foreground.

pub fn orphan() -> DeleteParams

Construct DeleteParams with PropagationPolicy::Orphan.

This orpans the dependents.

pub fn dry_run(self) -> DeleteParams

Perform a dryRun only

pub fn grace_period(self, secs: u32) -> DeleteParams

Set the duration in seconds before the object should be deleted.

pub fn preconditions(self, preconditions: Preconditions) -> DeleteParams

Set the condtions that must be fulfilled before a deletion is carried out.