Struct kube_client::api::Object

pub struct Object<P, U> where
    P: Clone,
    U: Clone,  {
    pub types: Option<TypeMeta>,
    pub metadata: ObjectMeta,
    pub spec: P,
    pub status: Option<U>,
}

A standard Kubernetes object with .spec and .status.

This is a convenience struct provided for serialization/deserialization. It is slightly stricter than [’DynamicObject`] in that it enforces the spec/status convention, and as such will not in general work with all api-discovered resources.

This can be used to tie existing resources to smaller, local struct variants to optimize for memory use. E.g. if you are only interested in a few fields, but you store tons of them in memory with reflectors.

Fields
types: Option<TypeMeta>

The type fields, not always present

metadata: ObjectMeta

Resource metadata

Contains information common to most resources about the Resource, including the object name, annotations, labels and more.

spec: P

The Spec struct of a resource. I.e. PodSpec, DeploymentSpec, etc.

This defines the desired state of the Resource as specified by the user.

status: Option<U>

The Status of a resource. I.e. PodStatus, DeploymentStatus, etc.

This publishes the state of the Resource as observed by the controller. Use U = NotUsed when a status does not exist.

Implementations

impl<P, U> Object<P, U> where
P: Clone,
U: Clone,

pub fn new(name: &str, ar: &ApiResource, spec: P) -> Object<P, U>

A constructor that takes Resource values from an ApiResource

pub fn within(self, ns: &str) -> Object<P, U>

Attach a namespace to an Object