Module kube_client::config

Kubernetes configuration objects from ~/.kube/config, $KUBECONFIG, or the cluster environment.

Usage

The Config has several constructors plus logic to infer environment.

Unless you have issues, prefer using Config::infer, and pass it to a Client.

Structs

AuthInfo

AuthInfo stores information to tell cluster who you are.

AuthProviderConfig

AuthProviderConfig stores auth for specified cloud provider.

Cluster

Cluster stores information to connect Kubernetes cluster.

Config

Configuration object detailing things like cluster URL, default namespace, root certificates, and timeouts.

Context

Context stores tuple of cluster and user information.

ExecAuthCluster

Cluster stores information to connect Kubernetes cluster used with auth plugins that have provideClusterInfo`` enabled. This is a copy of [kube::config::Cluster`] with certificate_authority passed as bytes without the path. Taken from clientauthentication/types.go#Cluster

ExecConfig

ExecConfig stores credential-plugin configuration.

InferConfigError

Failed to infer config

KubeConfigOptions

KubeConfigOptions stores options used when loading kubeconfig file.

Kubeconfig

Kubeconfig represents information on how to connect to a remote Kubernetes cluster

NamedAuthInfo

NamedAuthInfo associates name with authentication.

NamedCluster

NamedCluster associates name with cluster.

NamedContext

NamedContext associates name with context.

NamedExtension

NamedExtention associates name with extension.

Preferences

Preferences stores extensions for cli.

Enums

ExecInteractiveMode

ExecInteractiveMode define the interactity of the child process

InClusterError

Errors from loading in-cluster config

KubeconfigError

Possible errors when loading kubeconfig

LoadDataError

Errors from loading data from a base64 string or a file