Struct kube_client::config::Config

pub struct Config {
    pub cluster_url: Uri,
    pub default_namespace: String,
    pub root_cert: Option<Vec<Vec<u8>>>,
    pub connect_timeout: Option<Duration>,
    pub read_timeout: Option<Duration>,
    pub write_timeout: Option<Duration>,
    pub accept_invalid_certs: bool,
    pub auth_info: AuthInfo,
    pub proxy_url: Option<Uri>,
    pub tls_server_name: Option<String>,
}

Configuration object detailing things like cluster URL, default namespace, root certificates, and timeouts.

Usage

Construct a Config instance by using one of the many constructors.

Prefer Config::infer unless you have particular issues, and avoid manually managing the data in this struct unless you have particular needs. It exists to be consumed by the Client.

If you are looking to parse the kubeconfig found in a userโs home directory see Kubeconfig.

Fields
cluster_url: Uri

The configured cluster url

default_namespace: String

The configured default namespace

root_cert: Option<Vec<Vec<u8>>>

The configured root certificate

connect_timeout: Option<Duration>

Set the timeout for connecting to the Kubernetes API.

A value of None means no timeout

read_timeout: Option<Duration>

Set the timeout for the Kubernetes API response.

A value of None means no timeout

write_timeout: Option<Duration>

Set the timeout for the Kubernetes API request.

A value of None means no timeout

accept_invalid_certs: bool

Whether to accept invalid certificates

auth_info: AuthInfo

Stores information to tell the cluster who you are.

proxy_url: Option<Uri>

Optional proxy URL. Proxy support requires the socks5 feature.

tls_server_name: Option<String>

If set, apiserver certificate will be validated to contain this string

If not set, the cluster_url is used instead

Implementations

impl Config

pub fn new(cluster_url: Uri) -> Self

Construct a new config where only the cluster_url is set by the user. and everything else receives a default value.

Most likely you want to use Config::infer to infer the config from the environment.

pub async fn infer() -> Result<Self, InferConfigError>

Infer a Kubernetes client configuration.

First, a userโs kubeconfig is loaded from KUBECONFIG or ~/.kube/config. If that fails, an in-cluster config is loaded via Config::incluster. If inference from both sources fails, then an error is returned.

Config::apply_debug_overrides is used to augment the loaded configuration based on the environment.

pub fn incluster() -> Result<Self, InClusterError>

Load an in-cluster Kubernetes client configuration using Config::incluster_env.

pub fn incluster_env() -> Result<Self, InClusterError>

Load an in-cluster config using the KUBERNETES_SERVICE_HOST and KUBERNETES_SERVICE_PORT environment variables.

A service accountโs token must be available in /var/run/secrets/kubernetes.io/serviceaccount/.

This method matches the behavior of the official Kubernetes client libraries and is the default for both TLS stacks.

pub fn incluster_dns() -> Result<Self, InClusterError>

Load an in-cluster config using the API server at https://kubernetes.default.svc.

A service accountโs token must be available in /var/run/secrets/kubernetes.io/serviceaccount/.

This behavior does not match that of the official Kubernetes clients, but this approach is compatible with the rustls-tls feature without setting tls_server_name. See https://github.com/kube-rs/kube/issues/1003.

pub async fn from_kubeconfig( options: &KubeConfigOptions, ) -> Result<Self, KubeconfigError>

Create configuration from the default local config file

This will respect the $KUBECONFIG evar, but otherwise default to ~/.kube/config. You can also customize what context/cluster/user you want to use here, but it will default to the current-context.

pub async fn from_custom_kubeconfig( kubeconfig: Kubeconfig, options: &KubeConfigOptions, ) -> Result<Self, KubeconfigError>

Create configuration from a Kubeconfig struct

This bypasses kubeโs normal config parsing to obtain custom functionality.

pub fn apply_debug_overrides(&mut self)

Override configuration based on environment variables

This is only intended for use as a debugging aid, and the specific variables and their behaviour should not be considered stable across releases.

Currently, the following overrides are supported:

KUBE_RS_DEBUG_IMPERSONATE_USER: A Kubernetes user to impersonate, for example: system:serviceaccount:default:foo will impersonate the ServiceAccount foo in the Namespace default

KUBE_RS_DEBUG_IMPERSONATE_GROUP: A Kubernetes group to impersonate, multiple groups may be specified by separating them with commas

KUBE_RS_DEBUG_OVERRIDE_URL: A Kubernetes cluster URL to use rather than the one specified in the config, useful for proxying traffic through kubectl proxy