kube_client::error

Enum Error

pub enum Error {

    Api(ErrorResponse),
    HyperError(Error),
    Service(BoxError),
    FromUtf8(FromUtf8Error),
    LinesCodecMaxLineLengthExceeded,
    ReadEvents(Error),
    HttpError(Error),
    SerdeError(Error),
    BuildRequest(Error),
    InferConfig(InferConfigError),
    Discovery(DiscoveryError),
    RustlsTls(RustlsTlsError),
    TlsRequired,
    Auth(AuthError),

}

Possible errors when working with kube

Variants

Api(ErrorResponse)

ApiError for when things fail

This can be parsed into as an error handling fallback. Itโs also used in WatchEvent from watch calls.

Itโs quite common to get a 410 Gone when the resourceVersion is too old.

HyperError(Error)

Hyper error

Service(BoxError)

Service error

FromUtf8(FromUtf8Error)

UTF-8 Error

LinesCodecMaxLineLengthExceeded

Returned when failed to find a newline character within max length. Only returned by Client::request_events and this should never happen as the max is usize::MAX.

ReadEvents(Error)

Returned on std::io::Error when reading event stream.

HttpError(Error)

Http based error

SerdeError(Error)

Common error case when requesting parsing into own structs

BuildRequest(Error)

Failed to build request

InferConfig(InferConfigError)

Failed to infer config

Discovery(DiscoveryError)

Discovery errors

RustlsTls(RustlsTlsError)

Errors from Rustls TLS

TlsRequired

Missing TLS stacks when TLS is required

Auth(AuthError)

Errors related to client auth