Trait kube_client::ResourceExt

pub trait ResourceExt: Resource {
    fn name(&self) -> String;

    fn namespace(&self) -> Option<String>;

    fn resource_version(&self) -> Option<String>;

    fn uid(&self) -> Option<String>;

    fn labels(&self) -> &BTreeMap<String, String>;

    fn labels_mut(&mut self) -> &mut BTreeMap<String, String>;

    fn annotations(&self) -> &BTreeMap<String, String>;

    fn annotations_mut(&mut self) -> &mut BTreeMap<String, String>;

    fn owner_references(&self) -> &[OwnerReference];

    fn owner_references_mut(&mut self) -> &mut Vec<OwnerReference, Global>;

    fn finalizers(&self) -> &[String];

    fn finalizers_mut(&mut self) -> &mut Vec<String, Global>;
}

Helper methods for resources.

Required methods

fn name(&self) -> String

Returns the name of the resource, panicking if it is missing. Use this function if you know that name is set, for example when resource was received from the apiserver. Because of .metadata.generateName field, in other contexts name may be missing.

For non-panicking alternative, you can directly read name field on the self.meta().

fn namespace(&self) -> Option<String>

The namespace the resource is in

fn resource_version(&self) -> Option<String>

The resource version

fn uid(&self) -> Option<String>

Unique ID (if you delete resource and then create a new resource with the same name, it will have different ID)

fn labels(&self) -> &BTreeMap<String, String>

Returns resource labels

fn labels_mut(&mut self) -> &mut BTreeMap<String, String>

Provides mutable access to the labels

fn annotations(&self) -> &BTreeMap<String, String>

Returns resource annotations

fn annotations_mut(&mut self) -> &mut BTreeMap<String, String>

Provider mutable access to the annotations

fn owner_references(&self) -> &[OwnerReference]

Returns resource owner references

fn owner_references_mut(&mut self) -> &mut Vec<OwnerReference, Global>

Provides mutable access to the owner references

fn finalizers(&self) -> &[String]

Returns resource finalizers

fn finalizers_mut(&mut self) -> &mut Vec<String, Global>

Provides mutable access to the finalizers